Subject: Adding onClick functionality and scoring
Body:
Adding a function to handle the onClick when the image
is clicked and using the ID of the image to store in
state and then check to see whether the user has clicked
this image before.
Also, updating a currentScore and highScore state to keep
track of the scoring. Creating separate components and
lifting the state to the App so it can used throughout.